Move to @parcel/css by default
