--- packages/core/test-utils/src/.parcelrc-css
@@ -1,10 +1,0 @@
-{
-  "extends": "@parcel/config-default",
-  "transformers": {
-    "*.css": ["@parcel/transformer-postcss", "@parcel/transformer-css-experimental"]
-  },
-  "optimizers": {
-    "*.css": ["@parcel/optimizer-css"]
-  },
-  "reporters": []
-}
