--- packages/transformers/css-experimental/package.json
@@ -1,30 +1,0 @@
-{
-  "name": "@parcel/transformer-css-experimental",
-  "version": "2.3.2",
-  "license": "MIT",
-  "publishConfig": {
-    "access": "public"
-  },
-  "funding": {
-    "type": "opencollective",
-    "url": "https://opencollective.com/parcel"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/parcel-bundler/parcel.git"
-  },
-  "main": "lib/CSSTransformer.js",
-  "source": "src/CSSTransformer.js",
-  "engines": {
-    "node": ">= 12.0.0",
-    "parcel": "^2.3.2"
-  },
-  "dependencies": {
-    "@parcel/css": "^1.0.3",
-    "@parcel/plugin": "2.3.2",
-    "@parcel/source-map": "^2.0.0",
-    "@parcel/utils": "2.3.2",
-    "browserslist": "^4.6.6",
-    "nullthrows": "^1.1.1"
-  }
-}
