+++ packages/transformers/postcss/package.json
@@ -30,6 +30,7 @@
     "semver": "^5.7.1"
   },
   "devDependencies": {
-    "postcss": "^8.4.5"
+    "postcss": "^8.4.5",
+    "postcss-modules": "^4.3.1"
   }
 }
